fix(wp3): apply reviewer and security-audit findings
Pre-merge gate follow-ups (all non-blocking hardening/robustness; the embedding
product was already sound):

- Security S1 (highest-priority hardening): the ABI test pinned the three field
  values (pooling_type/attention_type/embeddings) but not the struct's trailing
  layout. Add a size_of assertion (160 bytes at b9726) so a future vendor-bump
  that reorders the tail after `embeddings` (samplers/n_samplers/ctx_other) is
  caught too, even though WP3 writes no tail field. Refresh on vendor-bump.

- Security S2: embedding_n_ctx narrowed usize->u32 before clamping (safe today
  only via the check_embed_fits precondition). Clamp before narrowing so
  truncation is impossible by construction.

- Reviewer F2: an empty string "" in the batch tokenizes to zero tokens on a
  no-BOS model (e.g. Qwen) and failed the whole call with a confusing engine-side
  error. Reject "" up front in R validation as rebirth_error_argument (fail-fast,
  predictable, model-independent), with a test and a clarified @param.

+++ rebirth/src/rust/rebirth-llm/src/embed.rs
@@ -250,7 +250,10 @@ impl LoadedModel {
     /// input length (at least 1) capped at the handle's context window. Sizing to
     /// the batch keeps the compute buffers small (the 16 GB rule, D-011).
     fn embedding_n_ctx(&self, longest: usize) -> u32 {
-        (longest.max(1) as u32).min(self.context_length())
+        // Clamp before narrowing so a `longest` beyond u32::MAX could never truncate
+        // to a wrong small value. Callers already run `check_embed_fits`, but this
+        // makes the narrowing safe by construction rather than by precondition.
+        longest.max(1).min(self.context_length() as usize) as u32
     }
 
     /// `RebirthError::Embed` if `len` tokens do not fit the context window, naming
